+++ Apps/Workbench/src/vite-env.d.ts
@@ -1,4 +1,4 @@
 /// <reference types="vite/client" />
 
-declare const __VOLITION_COMMIT__: string;
-declare const __VOLITION_VERSION__: string;
+declare const __WILLFORM_COMMIT__: string;
+declare const __WILLFORM_VERSION__: string;