+++ Bridges/Unreal/Source/WillformUnrealBridge/Private/WillformUnrealBridgeModule.cpp
@@ -1,0 +1,11 @@
+#include "WillformUnrealBridgeModule.h"
+
+void FWillformUnrealBridgeModule::StartupModule()
+{
+}
+
+void FWillformUnrealBridgeModule::ShutdownModule()
+{
+}
+
+IMPLEMENT_MODULE(FWillformUnrealBridgeModule, WillformUnrealBridge)