+++ Bridges/Unreal/Source/WillformUnrealEditor/Private/WillformUnrealEditorModule.cpp
@@ -1,0 +1,11 @@
+#include "WillformUnrealEditorModule.h"
+
+void FWillformUnrealEditorModule::StartupModule()
+{
+}
+
+void FWillformUnrealEditorModule::ShutdownModule()
+{
+}
+
+IMPLEMENT_MODULE(FWillformUnrealEditorModule, WillformUnrealEditor)